#34 Finished getting the selected modelName
