--- src/config/apiRoutes.ts
@@ -1,9 +1,0 @@
-/**
- * API routes for the application
- */
-export enum ApiRoute {
-    Login = '/api/login',
-    Logout = '/api/logout',
-    Register = '/api/register',
-    PdfToImage = '/api/pdfToImage',
-  }
