+++ src/config/constants.ts
@@ -55,3 +55,14 @@ export enum AppPages {
   Login = 'login',
   Register = 'register',
 }
+
+/**
+ * API routes for the application
+ */
+export enum ApiRoute {
+    Login = '/api/login',
+    Logout = '/api/logout',
+    Register = '/api/register',
+    PdfToImage = '/api/pdfToImage',
+    SelectedModels = '/api/selectedModels',
+  }
